Deleted old alembic revision file
