+++ backend/alembic/versions/ba231fd55e44_remove_userfooditem_table.py
@@ -1,0 +1,91 @@
+"""Remove UserFoodItem table
+
+Revision ID: ba231fd55e44
+Revises:
+Create Date: 2026-05-08 18:27:47.315644
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "ba231fd55e44"
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "recipes",
+        sa.Column(
+            "id", sa.Uuid(), server_default=sa.text("gen_random_uuid()"), nullable=False
+        ),
+        sa.Column("title", sa.String(length=50), nullable=False),
+        sa.Column("description", sa.String(length=600), nullable=True),
+        sa.Column("image_url", sa.Text(), nullable=True),
+        sa.Column("source_url", sa.Text(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_table(
+        "users",
+        sa.Column(
+            "id", sa.Uuid(), server_default=sa.text("gen_random_uuid()"), nullable=False
+        ),
+        sa.Column("email_address", sa.String(length=255), nullable=False),
+        sa.Column("password_hash", sa.String(length=255), nullable=False),
+        sa.Column("name", sa.String(length=30), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("email_address"),
+    )
+    op.create_table(
+        "food_items",
+        sa.Column(
+            "id", sa.Uuid(), server_default=sa.text("gen_random_uuid()"), nullable=False
+        ),
+        sa.Column("user_id", sa.Uuid(), nullable=False),
+        sa.Column("name", sa.String(length=50), nullable=False),
+        sa.Column("brand", sa.String(length=30), nullable=True),
+        sa.Column("barcode", sa.String(length=100), nullable=True),
+        sa.Column("category", sa.String(length=30), nullable=True),
+        sa.Column("image_url", sa.Text(), nullable=True),
+        sa.Column("quantity", sa.Numeric(precision=6, scale=2), nullable=False),
+        sa.Column("unit", sa.String(length=30), nullable=True),
+        sa.Column("expiration_date", sa.Date(), nullable=True),
+        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("user_id", "barcode", name="user_barcode"),
+    )
+    op.create_table(
+        "recipe_ingredients",
+        sa.Column(
+            "id", sa.Uuid(), server_default=sa.text("gen_random_uuid()"), nullable=False
+        ),
+        sa.Column("recipe_id", sa.Uuid(), nullable=False),
+        sa.Column("food_item_id", sa.Uuid(), nullable=False),
+        sa.Column("quantity", sa.Numeric(precision=10, scale=2), nullable=False),
+        sa.Column("unit", sa.String(length=30), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["food_item_id"],
+            ["food_items.id"],
+        ),
+        sa.ForeignKeyConstraint(["recipe_id"], ["recipes.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("recipe_ingredients")
+    op.drop_table("food_items")
+    op.drop_table("users")
+    op.drop_table("recipes")
+    # ### end Alembic commands ###
