--- backend/alembic/versions/f057b628bb83_initial_schemas.py
@@ -1,95 +1,0 @@
-"""Initial Schemas
-
-Revision ID: f057b628bb83
-Revises:
-Create Date: 2026-05-05 19:41:27.176147
-
-"""
-
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "f057b628bb83"
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "food_items",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("name", sa.String(length=50), nullable=False),
-        sa.Column("brand", sa.String(length=30), nullable=True),
-        sa.Column("barcode", sa.String(length=100), nullable=True),
-        sa.Column("category", sa.String(length=30), nullable=True),
-        sa.Column("image_url", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("barcode"),
-    )
-    op.create_table(
-        "recipes",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("title", sa.String(length=50), nullable=False),
-        sa.Column("description", sa.String(length=600), nullable=True),
-        sa.Column("image_url", sa.Text(), nullable=True),
-        sa.Column("source_url", sa.Text(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "users",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("email_address", sa.String(length=255), nullable=False),
-        sa.Column("password_hash", sa.String(length=255), nullable=False),
-        sa.Column("name", sa.String(length=30), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("email_address"),
-    )
-    op.create_table(
-        "recipe_ingredients",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("recipe_id", sa.Uuid(), nullable=False),
-        sa.Column("food_item_id", sa.Uuid(), nullable=False),
-        sa.Column("quantity", sa.Numeric(precision=10, scale=2), nullable=False),
-        sa.Column("unit", sa.String(length=30), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["food_item_id"],
-            ["food_items.id"],
-        ),
-        sa.ForeignKeyConstraint(["recipe_id"], ["recipes.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "user_food_items",
-        sa.Column("id", sa.Uuid(), nullable=False),
-        sa.Column("user_id", sa.Uuid(), nullable=False),
-        sa.Column("food_item_id", sa.Uuid(), nullable=False),
-        sa.Column("quantity", sa.Numeric(precision=6, scale=2), nullable=False),
-        sa.Column("unit", sa.String(length=30), nullable=True),
-        sa.Column("expiration_date", sa.Date(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["food_item_id"],
-            ["food_items.id"],
-        ),
-        sa.ForeignKeyConstraint(["user_id"], ["users.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("user_id", "food_item_id", name="user_food_id_uc"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("user_food_items")
-    op.drop_table("recipe_ingredients")
-    op.drop_table("users")
-    op.drop_table("recipes")
-    op.drop_table("food_items")
-    # ### end Alembic commands ###
